UserAuth Logic Improvments and UI enhancement (#247)
* commites

* final

* completed

* slight update design

* did some changes in structure

* env changed

* prettier fix

* connected to demo backend

* added comment

* Task Completed

* mobile scrolling done

* Image changed to vh

* test case

* preetier

* updated some things

* fix the lint

* fixed the login

* final changes done

* prettier fix

---------
