+++ context/LogContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+const LogContext = createContext();
+export default LogContext;
